chore: Refactor with Functions, part I
