+++ src/commands.js
@@ -1,26 +1,26 @@
-export { add } from './commands/add'
-export { branch } from './commands/branch'
-export { checkout } from './commands/checkout'
-export { clone } from './commands/clone'
-export { commit } from './commands/commit'
-export { config } from './commands/config'
-export { currentBranch } from './commands/currentBranch'
-export { fetch } from './commands/fetch'
-export { findRoot } from './commands/findRoot'
-export { getRemoteInfo } from './commands/getRemoteInfo'
-export { indexPack } from './commands/indexPack'
-export { init } from './commands/init'
-export { listBranches } from './commands/listBranches'
-export { listFiles } from './commands/listFiles'
-export { listTags } from './commands/listTags'
-export { log } from './commands/log'
-export { merge } from './commands/merge'
-export { pull } from './commands/pull'
-export { push } from './commands/push'
-export { readObject } from './commands/readObject'
-export { remove } from './commands/remove'
-export { resolveRef } from './commands/resolveRef'
-export { sign } from './commands/sign'
-export { status } from './commands/status'
-export { verify } from './commands/verify'
-export { version } from './commands/version'
+export * from './commands/add'
+export * from './commands/branch'
+export * from './commands/checkout'
+export * from './commands/clone'
+export * from './commands/commit'
+export * from './commands/config'
+export * from './commands/currentBranch'
+export * from './commands/fetch'
+export * from './commands/findRoot'
+export * from './commands/getRemoteInfo'
+export * from './commands/indexPack'
+export * from './commands/init'
+export * from './commands/listBranches'
+export * from './commands/listFiles'
+export * from './commands/listTags'
+export * from './commands/log'
+export * from './commands/merge'
+export * from './commands/pull'
+export * from './commands/push'
+export * from './commands/readObject'
+export * from './commands/remove'
+export * from './commands/resolveRef'
+export * from './commands/sign'
+export * from './commands/status'
+export * from './commands/verify'
+export * from './commands/version'
