+++ src/commands/checkout.js
@@ -2,6 +2,7 @@ import path from 'path'
 
 import { GitIndexManager, GitObjectManager, GitRefManager } from '../managers'
 import { FileSystem, GitCommit, GitTree } from '../models'
+import { writeTreeToDisk } from '../utils'
 
 import { config } from './config'
 
@@ -93,44 +94,5 @@ export async function checkout ({
   } catch (err) {
     err.caller = 'git.checkout'
     throw err
-  }
-}
-
-async function writeTreeToDisk ({ fs: _fs, dir, gitdir, index, prefix, tree }) {
-  const fs = new FileSystem(_fs)
-  for (let entry of tree) {
-    let { type, object } = await GitObjectManager.read({
-      fs,
-      gitdir,
-      oid: entry.oid
-    })
-    let entrypath = prefix === '' ? entry.path : `${prefix}/${entry.path}`
-    let filepath = path.join(dir, prefix, entry.path)
-    switch (type) {
-      case 'blob':
-        await fs.write(filepath, object)
-        let stats = await fs._lstat(filepath)
-        index.insert({
-          filepath: entrypath,
-          stats,
-          oid: entry.oid
-        })
-        break
-      case 'tree':
-        let tree = GitTree.from(object)
-        await writeTreeToDisk({
-          fs,
-          dir,
-          gitdir,
-          index,
-          prefix: entrypath,
-          tree
-        })
-        break
-      default:
-        throw new Error(
-          `Unexpected object type ${type} found in tree for '${entrypath}'`
-        )
-    }
   }
 }
