+++ src/commands/currentBranch.js
@@ -2,25 +2,7 @@ import path from 'path'
 
 import { GitRefManager } from '../managers'
 import { FileSystem } from '../models'
-
-// @see https://git-scm.com/docs/git-rev-parse.html#_specifying_revisions
-const regexs = [
-  new RegExp('refs/remotes/(.*)/HEAD'),
-  new RegExp('refs/remotes/(.*)'),
-  new RegExp('refs/heads/(.*)'),
-  new RegExp('refs/tags/(.*)'),
-  new RegExp('refs/(.*)')
-]
-
-function abbreviate (ref) {
-  for (const reg of regexs) {
-    let matches = reg.exec(ref)
-    if (matches) {
-      return matches[1]
-    }
-  }
-  return ref
-}
+import { abbreviateRef } from '../utils'
 
 /**
  * Get the name of the branch currently pointed to by .git/HEAD
@@ -42,7 +24,7 @@ export async function currentBranch ({
       depth: 2
     })
     if (fullname) return ref
-    return abbreviate(ref)
+    return abbreviateRef(ref)
   } catch (err) {
     err.caller = 'git.currentBranch'
     throw err
