+++ src/commands/listFiles.js
@@ -2,8 +2,8 @@ import path from 'path'
 
 import { GitIndexManager } from '../managers'
 import { FileSystem } from '../models'
+import { accumulateFilesFromOid } from '../utils'
 
-import { readObject } from './readObject'
 import { resolveRef } from './resolveRef'
 
 /**
@@ -38,25 +38,3 @@ export async function listFiles ({
     throw err
   }
 }
-
-async function accumulateFilesFromOid ({ gitdir, fs, oid, filenames, prefix }) {
-  const { object } = await readObject({ gitdir, fs, oid, filepath: '' })
-  // Note: this isn't parallelized because I'm too lazy to figure that out right now
-  for (const entry of object.entries) {
-    if (entry.type === 'tree') {
-      await accumulateFilesFromOid({
-        gitdir,
-        fs,
-        oid: entry.oid,
-        filenames,
-        prefix: posixJoin(prefix, entry.path)
-      })
-    } else {
-      filenames.push(posixJoin(prefix, entry.path))
-    }
-  }
-}
-
-// For some reason path.posix.join is undefined in webpack?
-const posixJoin = (prefix, filename) =>
-  prefix ? `${prefix}/${filename}` : filename
