+++ src/managers.js
@@ -1,9 +1,9 @@
-export { GitConfigManager } from './managers/GitConfigManager'
-export { GitIgnoreManager } from './managers/GitIgnoreManager'
-export { GitIndexManager } from './managers/GitIndexManager'
-export { GitObjectManager } from './managers/GitObjectManager'
-export { GitRefManager } from './managers/GitRefManager'
-export { GitRemoteConnection } from './managers/GitRemoteConnection'
-export { GitRemoteHTTP } from './managers/GitRemoteHTTP'
-export { GitRemoteManager } from './managers/GitRemoteManager'
-export { GitShallowManager } from './managers/GitShallowManager'
+export * from './managers/GitConfigManager'
+export * from './managers/GitIgnoreManager'
+export * from './managers/GitIndexManager'
+export * from './managers/GitObjectManager'
+export * from './managers/GitRefManager'
+export * from './managers/GitRemoteConnection'
+export * from './managers/GitRemoteHTTP'
+export * from './managers/GitRemoteManager'
+export * from './managers/GitShallowManager'
