+++ src/models.js
@@ -1,10 +1,11 @@
-export { FileSystem } from './models/FileSystem'
-export { GitCommit } from './models/GitCommit'
-export { GitConfig } from './models/GitConfig'
-export { GitIndex } from './models/GitIndex'
-export { GitObject } from './models/GitObject'
-export { GitPackIndex } from './models/GitPackIndex'
-export { GitPktLine } from './models/GitPktLine'
-export { GitRefSpec, GitRefSpecSet } from './models/GitRefSpec.js'
-export { GitTree } from './models/GitTree'
-export { SignedGitCommit } from './models/SignedGitCommit'
+export * from './models/FileSystem'
+export * from './models/GitCommit'
+export * from './models/GitConfig'
+export * from './models/GitIndex'
+export * from './models/GitObject'
+export * from './models/GitPackIndex'
+export * from './models/GitPktLine'
+export * from './models/GitRefSpec.js'
+export * from './models/GitRefSpecSet.js'
+export * from './models/GitTree'
+export * from './models/SignedGitCommit'
