+++ src/utils.js
@@ -1,8 +1,22 @@
-export { auth } from './utils/auth'
-export {
-  flatFileListToDirectoryStructure
-} from './utils/flatFileListToDirectoryStructure'
-export { log } from './utils/log'
-export { oauth2 } from './utils/oauth2'
-export { pkg } from './utils/pkg'
-export { sleep } from './utils/sleep'
+export * from './utils/abbreviateRef'
+export * from './utils/accumulateFilesFromOid'
+export * from './utils/auth'
+export * from './utils/cacheIsStale'
+export * from './utils/compareAge'
+export * from './utils/constructTree'
+export * from './utils/fetchPackfile'
+export * from './utils/findMergeBase'
+export * from './utils/flatFileListToDirectoryStructure'
+export * from './utils/getHeadTree'
+export * from './utils/getOidAtPath'
+export * from './utils/log'
+export * from './utils/listCommits'
+export * from './utils/listObjects'
+export * from './utils/oauth2'
+export * from './utils/pack'
+export * from './utils/pkg'
+export * from './utils/posixJoin'
+export * from './utils/resolveFile'
+export * from './utils/resolveTree'
+export * from './utils/sleep'
+export * from './utils/writeTreeToDisk'
