+++ src/utils/compareAge.js
@@ -1,0 +1,3 @@
+export function compareAge (a, b) {
+  return a.committer.timestamp - b.committer.timestamp
+}
